Move UserOptionValue
