+++ src/sentry/api/serializers/models/group.py
@@ -41,10 +41,10 @@ from sentry.models import (
     SentryAppInstallationToken,
     User,
     UserOption,
-    UserOptionValue,
 )
 from sentry.models.groupinbox import get_inbox_details
 from sentry.models.groupowner import get_owner_details
+from sentry.notifications.legacy_mappings import UserOptionValue
 from sentry.tagstore.snuba.backend import fix_tag_value_data
 from sentry.tsdb.snuba import SnubaTSDB
 from sentry.utils import snuba
