+++ src/sentry/mail/activity/release.py
@@ -1,8 +1,6 @@
 from collections import defaultdict
-from itertools import chain
-
-
 from django.db.models import Count, Q
+from itertools import chain
 
 from sentry.db.models.query import in_iexact
 from sentry.models import (
@@ -20,12 +18,12 @@ from sentry.models import (
     User,
     UserEmail,
     UserOption,
-    UserOptionValue,
 )
+from sentry.notifications.legacy_mappings import UserOptionValue
+from sentry.utils.compat import zip
 from sentry.utils.http import absolute_uri
 
 from .base import ActivityEmail
-from sentry.utils.compat import zip
 
 
 class ReleaseActivityEmail(ActivityEmail):
