+++ tests/snuba/api/serializers/test_group.py
@@ -19,8 +19,8 @@ from sentry.models import (
     GroupStatus,
     GroupSubscription,
     UserOption,
-    UserOptionValue,
 )
+from sentry.notifications.legacy_mappings import UserOptionValue
 from sentry.testutils import APITestCase, SnubaTestCase
 from sentry.testutils.helpers.datetime import iso_format, before_now
 from sentry.utils.compat import mock
